Better types for API

+++ UIImage+Spectrogram/AcquisitionContext.swift
@@ -10,23 +10,23 @@ import Foundation
 
 public class AcquisitionContext {
     
-    public var fftSize:Float
+    public var fftSize:Int
     public var sampleRate:Float
 
     public init(fftSize: Int, sampleRate: Float) {
-        self.fftSize = Float(fftSize)
+        self.fftSize = fftSize
         self.sampleRate = sampleRate
     }
     
-    public func frequencyStepForIndex(_ index:Float) -> Float {
-        return index*sampleRate / fftSize
+    public func frequencyStepForIndex(_ index:Int) -> Float {
+        return Float(index)*sampleRate / Float(fftSize)
     }
     
     public func indexForFrequency(_ targetFreq:Float) -> Int {
         var index = 0
         for i in 0...Int(fftSize) {
             index = i
-            let freq = frequencyStepForIndex(Float(i))
+            let freq = frequencyStepForIndex(i)
             if freq >= targetFreq {
                 return index
             }
@@ -35,7 +35,7 @@ public class AcquisitionContext {
     }
     
     public func durationInfosForSamples(_ samples:[Float]) -> (totalDuration:Float,chunkDuration:Float) {
-        let chunkDuration = (1.0/(sampleRate/fftSize))
+        let chunkDuration = (1.0/(sampleRate/Float(fftSize)))
         let totalDuration = Float(samples.count)/sampleRate
         return (totalDuration,chunkDuration)
     }
